backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
beautifulsoup4==4.12.2
lxml>=5.0.0
requests==2.31.0
tokenizers>=0.19.0
transformers>=4.40.0
torch>=2.6.0
sentence-transformers>=2.7.0
python-dotenv==1.0.0
pyreadline3>=3.5.0

